Made static texts dynamic

+++ app/Http/Controllers/Admin/LanguageController.php
@@ -87,6 +87,9 @@ class LanguageController extends Controller
     {
         try {
             $language = Language::findOrFail($id);
+            if ($language->lang == 'en') {
+                return response(['status' => 'error', 'message' => __('Cannot delete default language!')]);
+            }
             $language->delete();
             return response(['status' => 'success', 'message' => __('Deleted successfully!')]);
         } catch (\Throwable $th) {
